+ few basic helpful information for user + disable surya ocr - not implemented yet + place for logs (currently only pip error logs)

+++ run.sh
@@ -2,6 +2,10 @@
 
 DISABLE_VENV="${DISABLE_VENV:-0}"
 
+RED='\033[0;31m'
+CYAN='\033[0;36m'
+RESET='\033[0m'
+
 if [ "$DISABLE_VENV" -eq 1 ]; then
     echo "  .venv disabled"
 else
@@ -10,8 +14,25 @@ else
     source .venv/bin/activate
 fi
 
-pip install -r requirements.txt
-pip install -e .
+echo "Installing requirements..."
+if ! pip install -r requirements.txt 2>logs/init.log; then
+    echo "Failed to install dependencies from requirements.txt."
+    echo -e "Error log: $RED"
+    cat logs/init.log
+    echo -e "$RESET Please check the file and consider manually removing and recreating .venv if needed:"
+    echo -e "$CYAN    rm -rf .venv && python3 -m venv .venv && source .venv/bin/activate $RESET"
+    exit 1
+fi
+
+echo "Installing current package..."
+if ! pip install -e . 2>logs/init.log; then
+    echo "Failed to install the package in editable mode."
+    printf "Error log: %s" "$RED"
+    cat logs/init.log
+    echo -e "$RESET Please check the setup and consider manually removing and recreating .venv if needed:"
+    echo -e "$CYAN    rm -rf .venv && python3 -m venv .venv && source .venv/bin/activate $RESET"
+    exit 1
+fi
 
 if [ ! -f .env.localhost ]; then
   cp .env.localhost.example .env.localhost
@@ -29,7 +50,6 @@ echo "Pulling LLama3.2-vision model"
 ollama pull llama3.2-vision
 
 echo "Starting Redis"
-docker run -p 6379:6379 --restart always --detach redis &
 
 echo "Your ENV settings loaded from .env.localhost file: "
 printenv
@@ -37,17 +57,27 @@ printenv
 echo "Downloading models"
 #python -c 'from marker.models import load_all_models; load_all_models()'
 
-echo "Starting Celery worker"
+CELERY_BIN="$(pwd)/.venv/bin/celery"
+CELERY_PID=$(pgrep -f "$CELERY_BIN")
+REDIS_PORT=6379 # will move it to .envs in near future
+
+if lsof -i :$REDIS_PORT | grep LISTEN >/dev/null; then
+  echo "Redis is already running on port $REDIS_PORT. Skipping Redis start."
+else
+  echo "Starting Redis..."
+  docker run -p $REDIS_PORT:6379 --restart always --detach redis &
+fi
 
 
-echo "Starting FastAPI server"
+echo "Starting Celery Worker and FastAPI server"
 if [ $APP_ENV = 'production' ]; then
-    celery -A text_extract_api.celery_init worker --loglevel=info --pool=solo & # to scale by concurrent processing please run this line as many times as many concurrent processess you want to have running
+    celery -A text_extract_api.celery_init worker --loglevel=info --pool=solo & # to scale by concurrent processing please run this line as many times as many concurrent processess you want to have running; keep in mind that after next run they will be killed
     uvicorn text_extract_api.main:app --host 0.0.0.0 --port 8000;
 else
+
+  trap 'kill $(jobs -p) && exit' SIGINT SIGTERM
   (
-      trap 'kill 0' SIGINT;
-      celery -A text_extract_api.celery_app worker --loglevel=debug --pool=solo &
+      "$CELERY_BIN" -A text_extract_api.celery_app worker --loglevel=debug --pool=solo &
       uvicorn text_extract_api.main:app --host 0.0.0.0 --port 8000 --reload
   )
 fi
